requirements.txt
# Trading Journal — Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
itsdangerous>=2.0.0       # session middleware
python-multipart>=0.0.9   # form data parsing
